move submodule computationt to module_tree
