+++ crates/ra_hir/src/code_model_impl.rs
@@ -1,2 +1,2 @@
 mod krate; // `crate` is invalid ident :(
-pub(crate) mod module;
+mod module;
